task2/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0
Pillow==9.5.0
matplotlib==3.7.1
datasets==2.8.0
